chore: cleaned up some code to help reduce final package size
